fix(server): address /worktrees PR review issues (round 1)
- auth: restrict ?token= query-param fallback to /tasks/{id}/stream only
- auth: percent-decode token before constant-time comparison (fixes base64 tokens)
- web: surface useWorktrees() error in UI instead of showing empty state
- sse: replay round history for completed tasks instead of empty stream
- refactor: split http/mod.rs (was 1249 lines) into misc_routes.rs and sse_routes.rs
