+++ crates/harness-server/src/http/misc_routes.rs
@@ -1,0 +1,575 @@
+use axum::{
+    body::Bytes,
+    extract::{Path, State},
+    http::{HeaderMap, StatusCode},
+    response::{IntoResponse, Response},
+    Json,
+};
+use harness_protocol::methods::RpcRequest;
+use serde_json::json;
+use std::sync::Arc;
+
+use super::{state::AppState, task_routes};
+use crate::{router, task_runner};
+
+pub(crate) async fn health_check(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
+    let count = state.core.tasks.count();
+    let dirty = state.is_runtime_state_dirty();
+    let degraded = &state.degraded_subsystems;
+    let status = if degraded.is_empty() && !dirty {
+        "ok"
+    } else {
+        "degraded"
+    };
+    Json(json!({
+        "status": status,
+        "tasks": count,
+        "persistence": {
+            "degraded_subsystems": degraded,
+            "runtime_state_dirty": dirty,
+        }
+    }))
+}
+
+/// GET /projects/queue-stats — per-project queue stats alongside the global queue summary.
+pub(crate) async fn project_queue_stats(
+    State(state): State<Arc<AppState>>,
+) -> Json<serde_json::Value> {
+    let tq = &state.concurrency.task_queue;
+    let projects: serde_json::Map<String, serde_json::Value> = tq
+        .all_project_stats()
+        .into_iter()
+        .map(|(id, s)| {
+            (
+                id,
+                json!({
+                    "running": s.running,
+                    "queued": s.queued,
+                    "limit": s.limit,
+                }),
+            )
+        })
+        .collect();
+    Json(json!({
+        "global": {
+            "running": tq.running_count(),
+            "queued": tq.queued_count(),
+            "limit": tq.global_limit(),
+        },
+        "projects": projects,
+    }))
+}
+
+pub(crate) async fn handle_rpc(
+    State(state): State<Arc<AppState>>,
+    Json(req): Json<RpcRequest>,
+) -> Response {
+    match router::handle_request(&state, req).await {
+        Some(resp) => (StatusCode::OK, Json(resp)).into_response(),
+        None => StatusCode::NO_CONTENT.into_response(),
+    }
+}
+
+pub(crate) async fn github_webhook(
+    State(state): State<Arc<AppState>>,
+    headers: HeaderMap,
+    body: Bytes,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let secret = match state
+        .core
+        .server
+        .config
+        .server
+        .github_webhook_secret
+        .as_deref()
+    {
+        Some("") => {
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "invalid server.github_webhook_secret configuration"})),
+            )
+        }
+        Some(secret) => secret,
+        None => {
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "missing server.github_webhook_secret configuration"})),
+            )
+        }
+    };
+    let signature = match headers
+        .get("x-hub-signature-256")
+        .and_then(|value| value.to_str().ok())
+    {
+        Some(signature) => signature,
+        None => {
+            return (
+                StatusCode::UNAUTHORIZED,
+                Json(json!({"error": "missing header x-hub-signature-256"})),
+            )
+        }
+    };
+    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
+        return (
+            StatusCode::UNAUTHORIZED,
+            Json(json!({"error": "invalid webhook signature"})),
+        );
+    }
+
+    let event = match headers
+        .get("x-github-event")
+        .and_then(|value| value.to_str().ok())
+    {
+        Some(event) => event,
+        None => {
+            return (
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error": "missing header x-github-event"})),
+            )
+        }
+    };
+    if !crate::webhook::is_valid_github_event_name(event) {
+        return (
+            StatusCode::BAD_REQUEST,
+            Json(json!({"error": "invalid header x-github-event"})),
+        );
+    }
+
+    let (request, reason) =
+        match crate::webhook::parse_github_webhook_task_request(event, body.as_ref()) {
+            Ok(parsed) => parsed,
+            Err(error) => return (StatusCode::BAD_REQUEST, Json(json!({ "error": error }))),
+        };
+
+    let Some(mut req) = request else {
+        return (
+            StatusCode::OK,
+            Json(json!({
+                "status": "ignored",
+                "reason": reason,
+            })),
+        );
+    };
+
+    if req.project.is_none() {
+        req.project = Some(state.core.project_root.clone());
+    }
+
+    match task_routes::enqueue_task(&state, req).await {
+        Ok(task_id) => (
+            StatusCode::ACCEPTED,
+            Json(json!({
+                "status": "accepted",
+                "reason": reason,
+                "task_id": task_id.0,
+            })),
+        ),
+        Err(crate::services::execution::EnqueueTaskError::BadRequest(error)) => {
+            (StatusCode::BAD_REQUEST, Json(json!({ "error": error })))
+        }
+        Err(crate::services::execution::EnqueueTaskError::Internal(error)) => (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(json!({ "error": error })),
+        ),
+        Err(crate::services::execution::EnqueueTaskError::MaintenanceWindow {
+            retry_after_secs,
+        }) => (
+            StatusCode::SERVICE_UNAVAILABLE,
+            Json(json!({ "error": "maintenance_window", "retry_after": retry_after_secs })),
+        ),
+    }
+}
+
+pub(crate) async fn list_tasks(State(state): State<Arc<AppState>>) -> Response {
+    match state.core.tasks.list_all_summaries_with_terminal().await {
+        Ok(summaries) => Json(summaries).into_response(),
+        Err(e) => {
+            tracing::error!("list_tasks: database error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+pub(crate) async fn get_task(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    match state
+        .core
+        .tasks
+        .get_with_db_fallback(&harness_core::types::TaskId(id))
+        .await
+    {
+        Ok(Some(task)) => Json(task).into_response(),
+        Ok(None) => (
+            StatusCode::NOT_FOUND,
+            Json(json!({"error": "task not found"})),
+        )
+            .into_response(),
+        Err(e) => {
+            tracing::error!("get_task: database error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+/// GET /tasks/{id}/artifacts — all persisted artifacts for a task.
+pub(crate) async fn get_task_artifacts(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    let task_id = harness_core::types::TaskId(id);
+    match state.core.tasks.get_with_db_fallback(&task_id).await {
+        Ok(None) => {
+            return (
+                StatusCode::NOT_FOUND,
+                Json(json!({"error": "task not found"})),
+            )
+                .into_response();
+        }
+        Err(e) => {
+            tracing::error!("get_task_artifacts: database error: {e}");
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response();
+        }
+        Ok(Some(_)) => {}
+    }
+    match state.core.tasks.list_artifacts(&task_id).await {
+        Ok(artifacts) => Json(artifacts).into_response(),
+        Err(e) => {
+            tracing::error!("get_task_artifacts: list artifacts error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+/// GET /tasks/{id}/prompts — all persisted redacted prompts for a task.
+pub(crate) async fn get_task_prompts(
+    State(state): State<Arc<AppState>>,
+    Path(id): Path<String>,
+) -> Response {
+    let task_id = harness_core::types::TaskId(id);
+    match state.core.tasks.get_with_db_fallback(&task_id).await {
+        Ok(None) => {
+            return (
+                StatusCode::NOT_FOUND,
+                Json(json!({"error": "task not found"})),
+            )
+                .into_response();
+        }
+        Err(e) => {
+            tracing::error!("get_task_prompts: database error: {e}");
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response();
+        }
+        Ok(Some(_)) => {}
+    }
+    match state.core.tasks.get_prompts(&task_id).await {
+        Ok(prompts) => Json(prompts).into_response(),
+        Err(e) => {
+            tracing::error!("get_task_prompts: query error: {e}");
+            (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "internal server error"})),
+            )
+                .into_response()
+        }
+    }
+}
+
+/// GET /api/intake — current status of all intake channels and recent dispatches.
+pub(crate) async fn intake_status(State(state): State<Arc<AppState>>) -> Json<serde_json::Value> {
+    let intake_config = &state.core.server.config.intake;
+    let all_tasks = state.core.tasks.list_all();
+
+    let github_active: u64 = all_tasks
+        .iter()
+        .filter(|t| {
+            t.source.as_deref() == Some("github")
+                && !matches!(
+                    t.status,
+                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
+                )
+        })
+        .count() as u64;
+
+    let feishu_active: u64 = all_tasks
+        .iter()
+        .filter(|t| {
+            t.source.as_deref() == Some("feishu")
+                && !matches!(
+                    t.status,
+                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
+                )
+        })
+        .count() as u64;
+
+    let dashboard_active: u64 = all_tasks
+        .iter()
+        .filter(|t| {
+            (t.source.as_deref() == Some("dashboard") || t.source.is_none())
+                && !matches!(
+                    t.status,
+                    task_runner::TaskStatus::Done | task_runner::TaskStatus::Failed
+                )
+        })
+        .count() as u64;
+
+    let github_channel = json!({
+        "name": "github",
+        "enabled": intake_config.github.as_ref().map(|c| c.enabled).unwrap_or(false),
+        "repo": intake_config.github.as_ref().map(|c| c.repo.as_str()).unwrap_or(""),
+        "active": github_active,
+    });
+
+    let feishu_channel = json!({
+        "name": "feishu",
+        "enabled": state.intake.feishu_intake.is_some(),
+        "keyword": intake_config.feishu.as_ref().map(|c| c.trigger_keyword.as_str()).unwrap_or(""),
+        "active": feishu_active,
+    });
+
+    let dashboard_channel = json!({
+        "name": "dashboard",
+        "enabled": true,
+        "active": dashboard_active,
+    });
+
+    let mut recent_dispatches: Vec<serde_json::Value> = all_tasks
+        .iter()
+        .filter(|t| t.source.is_some())
+        .map(|t| {
+            json!({
+                "source": t.source,
+                "external_id": t.external_id,
+                "task_id": t.id.0,
+                "status": serde_json::to_value(&t.status).unwrap_or(json!("unknown")),
+                "pr_url": t.pr_url,
+            })
+        })
+        .collect();
+    recent_dispatches.truncate(10);
+
+    Json(json!({
+        "channels": [github_channel, feishu_channel, dashboard_channel],
+        "recent_dispatches": recent_dispatches,
+    }))
+}
+
+#[derive(serde::Deserialize)]
+pub(crate) struct IngestSignalRequest {
+    pub(crate) source: String,
+    #[serde(default)]
+    pub(crate) severity: Option<harness_core::types::Severity>,
+    pub(crate) payload: serde_json::Value,
+}
+
+/// Infer severity from a GitHub webhook payload: CI failure → High, changes_requested → Medium.
+pub(crate) fn infer_github_severity(
+    payload: &serde_json::Value,
+) -> Option<harness_core::types::Severity> {
+    if let Some(obj) = payload.as_object() {
+        if let (Some(action), Some(check_run)) = (
+            obj.get("action").and_then(|v| v.as_str()),
+            obj.get("check_run"),
+        ) {
+            if action == "completed" {
+                let conclusion = check_run
+                    .get("conclusion")
+                    .and_then(|v| v.as_str())
+                    .unwrap_or("");
+                if conclusion == "failure" {
+                    return Some(harness_core::types::Severity::High);
+                }
+            }
+        }
+        if let Some(review) = obj.get("review") {
+            let state = review.get("state").and_then(|v| v.as_str()).unwrap_or("");
+            if state.eq_ignore_ascii_case("changes_requested") {
+                return Some(harness_core::types::Severity::Medium);
+            }
+        }
+    }
+    None
+}
+
+/// POST /signals — ingest an external signal (CI failure, review feedback, etc.).
+///
+/// Validates the `x-hub-signature-256` HMAC-SHA256 header using the configured
+/// `server.github_webhook_secret`. Rate-limited to 100 requests per source per minute.
+pub(crate) async fn ingest_signal(
+    State(state): State<Arc<AppState>>,
+    headers: HeaderMap,
+    body: Bytes,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let secret = match state
+        .core
+        .server
+        .config
+        .server
+        .github_webhook_secret
+        .as_deref()
+    {
+        Some("") | None => {
+            return (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                Json(json!({"error": "server.github_webhook_secret not configured"})),
+            )
+        }
+        Some(s) => s,
+    };
+
+    let signature = match headers
+        .get("x-hub-signature-256")
+        .and_then(|v| v.to_str().ok())
+    {
+        Some(sig) => sig,
+        None => {
+            return (
+                StatusCode::UNAUTHORIZED,
+                Json(json!({"error": "missing header x-hub-signature-256"})),
+            )
+        }
+    };
+
+    if !crate::webhook::verify_github_signature(secret, signature, body.as_ref()) {
+        return (
+            StatusCode::UNAUTHORIZED,
+            Json(json!({"error": "invalid webhook signature"})),
+        );
+    }
+
+    let req: IngestSignalRequest = match serde_json::from_slice(&body) {
+        Ok(r) => r,
+        Err(e) => {
+            return (
+                StatusCode::BAD_REQUEST,
+                Json(json!({"error": format!("invalid payload: {e}")})),
+            )
+        }
+    };
+
+    if !state
+        .observability
+        .signal_rate_limiter
+        .check_and_increment(&req.source)
+    {
+        return (
+            StatusCode::TOO_MANY_REQUESTS,
+            Json(json!({"error": "rate limit exceeded: max 100 signals per minute per source"})),
+        );
+    }
+
+    let severity = req.severity.unwrap_or_else(|| {
+        if req.source == "github" {
+            infer_github_severity(&req.payload).unwrap_or(harness_core::types::Severity::Low)
+        } else {
+            harness_core::types::Severity::Low
+        }
+    });
+
+    let signal =
+        harness_core::types::ExternalSignal::new(req.source.clone(), severity, req.payload.clone());
+    let signal_id = signal.id.clone();
+
+    if let Err(e) = state.observability.events.log_external_signal(&signal) {
+        tracing::error!(source = %req.source, "failed to store external signal: {e}");
+        return (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(json!({"error": "failed to store signal"})),
+        );
+    }
+
+    tracing::info!(
+        source = %req.source,
+        severity = ?severity,
+        signal_id = %signal_id,
+        "external signal ingested"
+    );
+
+    (
+        StatusCode::OK,
+        Json(json!({"status": "accepted", "id": signal_id.as_str()})),
+    )
+}
+
+#[derive(serde::Deserialize)]
+pub(crate) struct PasswordResetRequest {
+    pub(crate) email: String,
+}
+
+/// POST /auth/reset-password — initiate a password reset.
+///
+/// Rate-limited per email address to prevent enumeration and brute-force.
+/// Always returns a generic success response regardless of whether the email
+/// exists, to avoid leaking account information.
+pub(crate) async fn password_reset(
+    State(state): State<Arc<AppState>>,
+    Json(req): Json<PasswordResetRequest>,
+) -> (StatusCode, Json<serde_json::Value>) {
+    let email = req.email.trim().to_lowercase();
+    if email.is_empty() {
+        return (
+            StatusCode::BAD_REQUEST,
+            Json(json!({"error": "email is required"})),
+        );
+    }
+
+    let limit = state
+        .core
+        .server
+        .config
+        .server
+        .password_reset_rate_limit_per_hour;
+
+    if !state
+        .observability
+        .password_reset_rate_limiter
+        .check_and_increment(&email)
+    {
+        return (
+            StatusCode::TOO_MANY_REQUESTS,
+            Json(json!({
+                "error": format!(
+                    "rate limit exceeded: max {} password reset requests per hour",
+                    limit
+                )
+            })),
+        );
+    }
+
+    tracing::info!(
+        email_hash = %format!("{:x}", {
+            use std::hash::{Hash, Hasher};
+            let mut h = std::collections::hash_map::DefaultHasher::new();
+            email.hash(&mut h);
+            h.finish()
+        }),
+        "password reset requested"
+    );
+
+    (
+        StatusCode::OK,
+        Json(
+            json!({"status": "ok", "message": "If that email is registered, a reset link has been sent."}),
+        ),
+    )
+}
